+++ examples/private_cluster/modules.tf
@@ -52,12 +52,14 @@ module "node_network_subnet" {
   resource_group_name  = module.rg.resource_group_name
   virtual_network_name = module.azure_virtual_network.virtual_network_name
 
+  name_suffix = "nodes"
+
   subnet_cidr_list = ["10.0.0.0/20"]
 
   service_endpoints = ["Microsoft.Storage"]
 }
 
-module "appgtw_network_subnet" {
+module "appgw_network_subnet" {
   source  = "claranet/subnet/azurerm"
   version = "x.x.x"
 
@@ -69,11 +71,13 @@ module "appgtw_network_subnet" {
   resource_group_name  = module.rg.resource_group_name
   virtual_network_name = module.azure_virtual_network.virtual_network_name
 
+  name_suffix = "appgw"
+
   subnet_cidr_list = ["10.0.20.0/24"]
 }
 
 module "global_run" {
-  source  = "claranet/run-common/azurerm"
+  source  = "claranet/run/azurerm"
   version = "x.x.x"
 
   client_name    = var.client_name
@@ -84,10 +88,13 @@ module "global_run" {
 
   monitoring_function_splunk_token = var.monitoring_function_splunk_token
 
-
   resource_group_name = module.rg.resource_group_name
 
   tenant_id = var.azure_tenant_id
+}
+
+resource "tls_private_key" "key" {
+  algorithm = "RSA"
 }
 
 module "aks" {
@@ -113,7 +120,7 @@ module "aks" {
   private_dns_zone_id     = azurerm_private_dns_zone.private_dns_zone.id
 
   agic_enabled           = true
-  appgw_subnet_id        = module.appgtw_network_subnet.subnet_id
+  appgw_subnet_id        = module.appgw_network_subnet.subnet_id
   appgw_identity_enabled = true
 
   default_node_pool = {
@@ -139,8 +146,8 @@ module "aks" {
   ]
 
   linux_profile = {
-    username = "user"
-    ssh_key  = "ssh_priv_key"
+    username = "nodeadmin"
+    ssh_key  = tls_private_key.key.public_key_openssh
   }
 
   oms_log_analytics_workspace_id = module.global_run.log_analytics_workspace_id